Develop

+++ handlers/answer_handler/answer_handler.go
@@ -11,41 +11,49 @@ import (
 )
 
 // DeleteAnswer
-// @Tags answer_handler
+// @Tags answer
 // @Accept json
 // @Produce json
 // @Param Authorization header string true "Bearer Token"
 // @Success 201 {object} models.Answer
-// @Router /api/answer_handler/delete [delete]
+// @Router /api/answer/delete [delete]
 func DeleteAnswer(c *gin.Context) {
 	var answer models.Answer
 	if err := c.ShouldBindJSON(&answer); err != nil {
 		c.IndentedJSON(http.StatusBadRequest, gin.H{"message": "Invalid JSON format"})
 		return
 	}
-	//todo : check if user has permission to delete answer_handler
 
-	result := database.DB.Delete(&answer)
-	if result.Error != nil {
-		c.IndentedJSON(http.StatusInternalServerError, gin.H{"message": "Error creating answer_handler"})
+	user := helper.FetchUserFromToken(c.GetHeader("Authorization"))
+	if user == nil {
+		c.IndentedJSON(http.StatusInternalServerError, gin.H{"message": "failed to fetch user data"})
 		return
 	}
-	c.IndentedJSON(http.StatusCreated, answer)
+
+	if user.IsAdmin || user.UserId == answer.UserId {
+		result := database.DB.Delete(&answer)
+		if result.Error != nil {
+			c.IndentedJSON(http.StatusInternalServerError, gin.H{"message": "Error deleting answer"})
+			return
+		}
+		c.IndentedJSON(http.StatusCreated, answer)
+	} else {
+		c.IndentedJSON(http.StatusUnauthorized, gin.H{"message": "you can only delete your own answers"})
+	}
 }
 
 // CorrectAnswer
-// @Tags answer_handler
+// @Tags answer
 // @Accept json
 // @Produce json
 // @Param Authorization header string true "Bearer Token"
 // @Success 201 {object} models.Answer
-// @Router /api/answer_handler/correctAnswer/{id} [get]
+// @Router /api/answer/correctAnswer/{id} [get]
 func CorrectAnswer(c *gin.Context) {
-	//todo : check if user has asked the question_handler to mark it as correct
 	id, _ := strconv.ParseInt(c.Param("id"), 10, 64)
 	var answer models.Answer
 	if err := database.DB.Where("answer_id = ?", id).First(&answer).Error; err != nil {
-		c.IndentedJSON(http.StatusBadRequest, gin.H{"message": "answer_handler not found"})
+		c.IndentedJSON(http.StatusBadRequest, gin.H{"message": "answer not found"})
 		return
 	}
 
@@ -64,24 +72,24 @@ func CorrectAnswer(c *gin.Context) {
 	if user.UserId == question.UserId {
 		answer.IsCorrectAnswer = true
 		if updateError := database.DB.Save(&answer).Error; updateError != nil {
-			c.IndentedJSON(http.StatusBadRequest, gin.H{"message": "failed to update the answer_handler"})
+			c.IndentedJSON(http.StatusBadRequest, gin.H{"message": "failed to update the answer"})
 			return
 		}
 		c.IndentedJSON(http.StatusOK, answer)
 	} else {
-		c.IndentedJSON(http.StatusBadRequest, gin.H{"message": "only the user asked the question, may mark it as correct"})
+		c.IndentedJSON(http.StatusUnauthorized, gin.H{"message": "only the user asked the question, may mark it as correct"})
 	}
 
 }
 
 // AddAnswer
-// @Tags answer_handler
+// @Tags answer
 // @Accept json
 // @Produce json
 // @Param Authorization header string true "Bearer Token"
-// @Param answer_handler body models.Answer true "Answer object"
+// @Param answer body models.Answer true "Answer object"
 // @Success 201 {object} models.Answer
-// @Router /api/answer_handler/add [post]
+// @Router /api/answer/add [post]
 func AddAnswer(c *gin.Context) {
 	var answer models.Answer
 	if err := c.ShouldBindJSON(&answer); err != nil {
@@ -114,7 +122,7 @@ func AddAnswer(c *gin.Context) {
 
 	result := database.DB.Create(&answer)
 	if result.Error != nil {
-		c.IndentedJSON(http.StatusInternalServerError, gin.H{"message": "Error creating answer_handler"})
+		c.IndentedJSON(http.StatusInternalServerError, gin.H{"message": "Error creating answer"})
 		return
 	}
 	c.IndentedJSON(http.StatusCreated, answer)
